refactor <meta> search & parse
This commit refactors the while loop used to search and parse <meta> elements. Nested for loops iterate through the keys of the tags object and search for data matching each of the items in the value's array. When content for a key is found it is pushed to the META object and subsequent content for that key is ignored.

closes-issues: #4, #7, #8